+++ src/main/java/racingcar/domain/NumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 public interface NumberGenerator {
     int generateNumber(int bound);
