+++ src/main/java/racingcar/domain/Referee.java
@@ -1,7 +1,5 @@
-package racingcar.model;
+package racingcar.domain;
 
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
 public class Referee {
